+++ package.json
@@ -12,6 +12,7 @@
 		"lint-fix": "prettier --write \"packages/*/src/**/*.{ts,tsx}\"",
 		"clean": "find . -name \"node_modules\" -exec rm -rf '{}' + && find . -name \"lib\" -exec rm -rf '{}' +",
 		"format": "pretty-quick --pattern \"./packages/**/*.*(js|ts|tsx)\"",
+		"bundle": "npx lerna exec --parallel -- yarn bundle",
 		"docz:dev": "docz dev",
 		"docz:build": "docz build",
 		"docz:serve": "docz build && docz serve",
@@ -25,19 +26,19 @@
 		"@tpr/theming": "file:./packages/theming"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.9.0",
-		"@babel/preset-env": "^7.9.5",
-		"@rollup/plugin-commonjs": "^11.1.0",
-		"@rollup/plugin-node-resolve": "^7.1.3",
-		"@testing-library/jest-dom": "^5.5.0",
-		"@testing-library/react": "^10.0.2",
+		"@babel/core": "^7.10.2",
+		"@babel/preset-env": "^7.10.2",
+		"@rollup/plugin-commonjs": "^12.0.0",
+		"@rollup/plugin-node-resolve": "^8.0.0",
+		"@testing-library/jest-dom": "^5.9.0",
+		"@testing-library/react": "^10.0.5",
 		"@testing-library/react-hooks": "^3.2.1",
 		"@types/jest": "^25.2.1",
 		"@types/react": "^16.9.34",
 		"@types/react-dom": "^16.9.6",
 		"@xstate/react": "^0.8.1",
 		"autoprefixer": "^9.7.6",
-		"babel-jest": "^25.3.0",
+		"babel-jest": "^26.0.1",
 		"babel-loader": "^8.1.0",
 		"codecov": "^3.6.1",
 		"copyfiles": "^2.2.0",
@@ -46,8 +47,8 @@
 		"gatsby-plugin-sass": "^2.3.0",
 		"husky": "^4.2.5",
 		"identity-obj-proxy": "^3.0.0",
-		"immutability-helper": "^3.0.2",
-		"jest": "^25.3.0",
+		"immutability-helper": "^3.1.1",
+		"jest": "^26.0.1",
 		"lerna": "^3.19.0",
 		"lerna-changelog": "^1.0.0",
 		"node-sass": "^4.14.0",
@@ -59,22 +60,21 @@
 		"react-dom": "^16.13.1",
 		"react-test-renderer": "^16.13.1",
 		"rimraf": "^3.0.0",
-		"rollup": "^2.6.1",
-		"rollup-plugin-invariant": "^0.5.6",
+		"rollup": "^2.12.0",
 		"rollup-plugin-local-resolve": "^1.0.7",
 		"rollup-plugin-peer-deps-external": "^2.2.2",
-		"rollup-plugin-postcss": "^2.8.2",
-		"rollup-plugin-sourcemaps": "^0.5.0",
-		"rollup-plugin-terser": "^5.3.0",
-		"rollup-plugin-typescript2": "^0.27.0",
-		"ts-jest": "^25.3.1",
-		"ts-loader": "^7.0.0",
-		"ts-node": "^8.8.2",
-		"tslib": "^1.10.0",
-		"typescript": "^3.8.3"
+		"rollup-plugin-postcss": "^3.1.1",
+		"rollup-plugin-sourcemaps": "^0.6.2",
+		"rollup-plugin-terser": "^6.1.0",
+		"rollup-plugin-typescript2": "^0.27.1",
+		"ts-jest": "^26.1.0",
+		"ts-loader": "^7.0.5",
+		"ts-node": "^8.10.2",
+		"tslib": "^2.0.0",
+		"typescript": "^3.9.3"
 	},
 	"resolutions": {
-		"@types/react": "16.9.34"
+		"@types/react": "16.9.35"
 	},
 	"lint-staged": {
 		"*.{ts,tsx}": [
