+++ packages/forms/package.json
@@ -39,22 +39,21 @@
 		"@tpr/theming": "file:../theming",
 		"@types/lodash.merge": "^4.6.6",
 		"@types/match-sorter": "^4.0.0",
-		"@types/qs": "^6.9.2",
-		"date-fns": "^2.13.0",
-		"downshift": "^5.3.0",
-		"final-form": "^4.19.1",
+		"@types/qs": "^6.9.3",
+		"date-fns": "^2.14.0",
+		"downshift": "^5.4.2",
+		"final-form": "^4.20.0",
 		"lodash.merge": "^4.6.2",
 		"match-sorter": "^4.1.0",
 		"qs": "^6.9.4",
-		"react-final-form": "^6.4.0",
+		"react-final-form": "^6.5.0",
 		"tslib": "^2.0.0"
 	},
 	"publishConfig": {
 		"access": "public"
 	},
 	"devDependencies": {
-		"cross-env": "^7.0.0",
-		"typescript": "^3.8.3"
+		"cross-env": "^7.0.2"
 	},
 	"gitHead": "57aec9c9de22596d6343bae07eb6b198322a0bee"
 }
