+++ packages/icons/package.json
@@ -36,12 +36,11 @@
 		"access": "public"
 	},
 	"devDependencies": {
-		"@tpr/core": "file:../core",
-		"cross-env": "^7.0.0",
-		"typescript": "^3.8.3"
+		"cross-env": "^7.0.0"
 	},
 	"dependencies": {
-		"tslib": "^1.11.1"
+		"@tpr/core": "file:../core",
+		"tslib": "^2.0.0"
 	},
 	"gitHead": "57aec9c9de22596d6343bae07eb6b198322a0bee"
 }
