+++ packages/layout/package.json
@@ -42,15 +42,14 @@
 		"@tpr/icons": "file:../icons",
 		"@tpr/theming": "file:../theming",
 		"@xstate/react": "^0.8.1",
-		"final-form": "^4.18.7",
-		"qs": "^6.9.1",
-		"react-final-form": "^6.3.5",
-		"tslib": "^1.11.1",
-		"xstate": "^4.9.0"
+		"final-form": "^4.20.0",
+		"qs": "^6.9.4",
+		"react-final-form": "^6.5.0",
+		"tslib": "^2.0.0",
+		"xstate": "^4.10.0"
 	},
 	"devDependencies": {
-		"cross-env": "^7.0.0",
-		"typescript": "^3.8.3"
+		"cross-env": "^7.0.2"
 	},
 	"publishConfig": {
 		"access": "public"
